fix(web-device): drop two redundant `NOW as i64` casts in test fixtures
`NOW` is already `i64` (`const NOW: i64` at lib.rs:841), so both casts were
no-ops that `clippy::unnecessary_cast` denies.

WHY THIS REACHED CI RATHER THAN MY MACHINE, which is the part worth recording.
I ran `cargo clippy --all-features -- -D warnings`, which checks the library
only. Both casts are in `#[cfg(test)]` code, and CI checks test targets too —
its error names the unit as `selfsame-web-device (lib test)`.

So my local command was strictly weaker than the one it was standing in for,
and a clean local run meant less than I took it to mean. The command that
actually matches CI is `cargo clippy --all-features --all-targets --
-D warnings`, which is now clean here.

+++ crates/selfsame-web-device/src/lib.rs
@@ -1082,7 +1082,7 @@ mod tests {
                 "revoked_credential_ids": issuer.revoked_credential_ids,
                 // A fetch stamp, not an age. The fixture's issuer carries an age,
                 // so the equivalent stamp is that far before `now`.
-                "fetched_at_seconds": NOW as i64 - issuer.closure_age_seconds,
+                "fetched_at_seconds": NOW - issuer.closure_age_seconds,
                 "also_known_as": issuer.also_known_as,
             })
         };
@@ -1329,7 +1329,7 @@ mod tests {
             assertion_methods: vec![method(byte)],
             revoked_credential_ids: vec![],
             also_known_as: fixture.issuer.also_known_as.clone(),
-            fetched_at_seconds: NOW as i64,
+            fetched_at_seconds: NOW,
         };
 
         let agreed = agree_closures(&fixture.profile, &[observation("app-own", 1), observation("state-1", 1)]);
